addressed code reviews
