+++ newton/newton-api.c
@@ -51,6 +51,7 @@
 #include "version.h"
 #include "common-timeStamps.h"
 #include "data-structures.h"
+#include "newton-data-structures.h"
 
 #include "newton-parser.h"
 #include "newton-lexer.h"
