+++ newton/newton-data-structures.h
@@ -1,5 +1,5 @@
 /*
-	Authored 2017. Jonathan Lim
+	Authored 2016. Jonathan Lim.
 
 	All rights reserved.
 
@@ -35,60 +35,21 @@
 	POSSIBILITY OF SUCH DAMAGE.
 */
 
-#include <math.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <stdbool.h>
-#include <sys/time.h>
-#include <getopt.h>
-#include <setjmp.h>
-#include <stdint.h>
-#include "flextypes.h"
-#include "flexerror.h"
-#include "flex.h"
-#include "common-errors.h"
-#include "version.h"
-#include "common-timeStamps.h"
-#include "data-structures.h"
+typedef struct NewtonAPIReport NewtonAPIReport;
+typedef struct ConstraintReport ConstraintReport;
 
-#include "newton-parser.h"
-#include "newton-lexer.h"
-#include "newton-symbolTable.h"
-#include "newton.h"
-#include "newton-irPass-dotBackend.h"
-
-extern char* gNewtonAstNodeStrings[kNoisyIrNodeTypeMax];
-
-
-void		
-processNewtonFile(NoisyState *  N, char *  filename)
+struct ConstraintReport
 {
+  bool satisfiesDimensionConstraint;
+  char dimensionErrorMessage[1024];
+  bool satisfiesValueConstraint;
+  char valueErrorMessage[1024];
 
-	/*
-	 *	Tokenize input, then parse it and build AST + symbol table.
-	 */
-	newtonLexInit(N, filename);
+  ConstraintReport* next;
+};
 
-	/*
-	 *	Create a top-level scope, then parse.
-	 */
-	N->newtonIrTopScope = newtonSymbolTableAllocScope(N);
-	N->newtonIrRoot = newtonParse(N, N->newtonIrTopScope);
-
-	/*
-	 *	Dot backend.
-	 */
-	if (N->irBackends & kNoisyIrBackendDot)
-    fprintf(stdout, "%s\n", irPassDotBackend(N, N->newtonIrTopScope, N->newtonIrRoot, gNewtonAstNodeStrings));
-
-
-
-	// if (N->mode & kNoisyConfigModeCallTracing)
-	// {
-	// 	noisyConfigTimeStampDumpTimeline(N);
-	// }
-
-    noisyConsolePrintBuffers(N);
-}
+struct NewtonAPIReport
+{
+  ConstraintReport* firstConstraintReport;
+};
