+++ tests/test-common.c
@@ -47,6 +47,7 @@
 
 #include "common-timeStamps.h"
 #include "data-structures.h"
+#include "newton-data-structures.h"
 #include "common-irHelpers.h"
 #include "common-lexers-helpers.h"
 #include "newton-parser.h"
